requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
APScheduler==3.10.4
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
